+++ es/utils.js
@@ -682,3 +682,42 @@ export function print_opts(opts){
 		$("#pedigree_data").append("<span>"+key + ":" + opts[key]+"; </span>");
 	}
 }
+
+export function is_fullscreen(){
+	return (document.fullscreenElement || document.mozFullScreenElement || document.webkitFullscreenElement || document.msFullscreenElement);
+}
+
+
+export function get_svg_dimensions(opts) {
+	return {'width' : (is_fullscreen()? window.innerWidth  : opts.width),
+			'height': (is_fullscreen()? window.innerHeight : opts.height)};
+}
+
+export function get_tree_dimensions(opts) {
+	// / get score at each depth used to adjust node separation
+	let svg_dimensions = get_svg_dimensions(opts);
+	let maxscore = 0;
+	let generation = {};
+	for(let i=0; i<opts.dataset.length; i++) {
+		let depth = getDepth(opts.dataset, opts.dataset[i].name);
+		let children = getAllChildren(opts.dataset, opts.dataset[i]);
+
+		// score based on no. of children and if parent defined
+		let score = 1 + (children.length > 0 ? 0.55+(children.length*0.25) : 0) + (opts.dataset[i].father ? 0.25 : 0);
+		if(depth in generation)
+			generation[depth] += score;
+		else
+			generation[depth] = score;
+
+		if(generation[depth] > maxscore)
+			maxscore = generation[depth];
+	}
+
+	let max_depth = Object.keys(generation).length*opts.symbol_size*3.5;
+	let tree_width =  (svg_dimensions.width - opts.symbol_size > maxscore*opts.symbol_size*1.65 ?
+					   svg_dimensions.width - opts.symbol_size : maxscore*opts.symbol_size*1.65);
+	let tree_height = (svg_dimensions.height - opts.symbol_size > max_depth ?
+					   svg_dimensions.height - opts.symbol_size : max_depth);
+	return {'width': tree_width, 'height': tree_height};
+}
+
